# Improve compilation times of pipeline apps